Change behavior for armor < 10
I misunderstood how the Weapon::hit() function was working.
I originally thought that hit() and hit(armor) were the same,
so I designed LightningStaff to only react to hit(armor).

I took out the code from hit(armor) that was checking if
the armor was == 0 since Weapon::hit() (with an implied
armor value of 0) doesn't even run that code.